+++ src/Get-First_function.ps1
@@ -71,7 +71,7 @@ function Get-First
     [bool] $isFirstItem = $True
     [string] $oldVal = $Null
     [string] $newVal = $Null
-    foreach ( $obj in $input ){
+    foreach ( $obj in @($input | Select-Object *) ){
         [string] $propKeyStr = ''
         foreach ($p in $Property){
             $propKeyStr += $obj.$p